reorganize types files
